changed in db regarding autoindexing requested by search team
